+++ libft/libft.h
@@ -93,10 +93,10 @@ char				*ft_strrev(char *str);
 long				ft_power(long n, long pow);
 char				*ft_itoa_negative(intmax_t v, int b);
 char				*ft_itoa_negative_decimal(char *s);
-char				*ft_itoa_double(double nbr, int k);
+char				*ft_itoa_double(long double nbr, int k);
 long				ft_roundnbr(long n, int i);
-char				*ft_itoa_exp(double nbr, int k);
-char				*ft_itoa_double_g(double nbr, int k);
+char				*ft_itoa_exp(long double nbr, int k);
+char				*ft_itoa_double_g(long double nbr, int k);
 char				*ft_itoa_signed(intmax_t num);
 char				*ft_itoa_unsigned(uintmax_t num, int base);
